Updated MefareshParser: Changed the design, added abarbanel parsing
MefareshParser is no longer an abstract class. It gets the mefaresh name from the first line of the raw file. To add a new mefaresh, the class just needs it to be added to the names Map. See class for more info.

Deleted RambanParser and RambanParserTest classes.

Added extensive MefareshParserTest class to test both ramban and abarbanel

Updated config file with "tanachmefarshim" folder and added a method to TestUtils that lets me set up a parser that has different files in the same directory

issue #51
issue #50