Updated MefareshParser: Added 9 new Mefarshim
MefareshParser now relies on the line that has the name of the book to know which book its in. The line has to be accurate to the JbsUtils book names, and have *only* the book name (and not "ספר " before). This update changes the BEGIN_SEFER condition - its true if the line is a known book.

Updated getCleanLine - added the library Jsoup to the Maven dependencies, this library correctly parses the HTML tags.

Added 9 new Mefarshim to the createMefareshMap method

Updated the MefareshParserTest class, different mefarshim have different test methods.

Updated Parser.java: on EOF, the jsonObject will flush only if it has a key of type JBO_TEXT or JBO_TEXT_NIKUD. This is because some files end with an empty verse and this is unwanted. The behavior stays as expected on all the other files (didn't break anything)

closed #51
issue #52