Updated MefareshParser: Updates to fit all the new Mefarshim files
This commit is relevant to the jbs-data commit with the new Python script and 51 mefarshim files.
Updated the parser to be more general and support all of them.
Specifically: BEGIN_ENG_TITLE is a new important match type, it is supposed to be the second line of the raw file and will contain 'perush *name*', where the name is the English name of the Mefaresh, example: 'perush rashi'. This replaces the mefareshMap which required knowing in the code the names of all Mefarshim. This new way is more modular

Test class also updated

issue #52